feat(336): renaming token folders to kebab-case
