+++ src/lib/components/00-tokens/line-height/line-height.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import LineHeightDisplay from './lineHeight';
+import LineHeightDisplay from './line-height';
 
 const leadingsVariants = {
   'leading-none': '1',
